@@ -4,12 +4,12 @@
   "description": "MaarchCourrier",
   "scripts": {
     "ng": "ng",
-    "ng-update": "ng update --allow-dirty",
-    "build-prod": "ng build --configuration production",
-    "build-watch": "ng build --watch",
+    "ng-update": "ng update --allow-dirty --force",
+    "build-prod": "ng build --project \"maarchCourrier\" --configuration production",
+    "build-watch": "ng build --project \"maarchCourrier\" --watch",
     "build-prod-addin-outlook": "ng build --project \"addin-outlook\" --configuration production",
     "build-addin-outlook": "ng build --project \"addin-outlook\"",
-    "build": "ng build",
+    "build": "ng build --project \"maarchCourrier\"",
     "build-release": "npm run build-prod && npm run build-prod-addin-outlook && npm run reload-packages-prod && npm run reload-packages-prod-addin-outlook",
     "check-update": "ncu",
     "dep-list": "npm list --depth=0",
@@ -20,8 +20,8 @@
     "reload-packages-addin-outlook": "cd plugins/addin-outlook && rm -Rf node_modules && rm -f package-lock.json && npm cache clean --force && && npm config set legacy-peer-deps true && npm i",
     "reload-packages-prod": "rm -Rf node_modules && rm package-lock.json && npm cache clean --force && npm config set legacy-peer-deps && npm i --production",
     "reload-packages-prod-addin-outlook": "cd plugins/addin-outlook && rm -Rf node_modules && rm package-lock.json && npm cache clean --force && npm config set legacy-peer-deps && npm i --production",
-    "create-component": "echo Component name ? && read varname && ng g c $varname --module app --style=scss",
-    "create-component-admin": "echo Component name ? && read varname && ng g c administration/$varname --module administration --style=scss",
+    "create-component": "echo Component name ? && read varname && ng g c $varname --module app --style=scss",
+    "create-component-admin": "echo Component name ? && read varname && ng g c administration/$varname --module administration --style=scss",
     "cypress:open": "cypress open --config-file test/e2e/cypress-config.json",
     "cypress:run": "cypress run --browser firefox --headless --config-file test/e2e/cypress-config.json"
   },
@@ -34,25 +34,25 @@
     "core-js": "^3.15.1",
     "tinymce": "^5.8.2",
     "tinymce-i18n": "^20.12.25",
-    "zone.js": "~0.11.4",
-    "tslib": "^2.3.1"
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.6",
-    "@angular-eslint/eslint-plugin": "^2.1.1",
-    "@angular/animations": "^13.2.6",
-    "@angular/cdk": "^13.2.6",
-    "@angular/cli": "^13.2.6",
-    "@angular/common": "^13.2.6",
-    "@angular/compiler": "^13.2.6",
-    "@angular/compiler-cli": "^13.2.6",
-    "@angular/core": "^13.2.6",
-    "@angular/forms": "^13.2.6",
-    "@angular/material": "^13.3.0",
-    "@angular/platform-browser": "^13.2.6",
-    "@angular/platform-browser-dynamic": "^13.2.6",
-    "@angular/platform-server": "^13.2.6",
-    "@angular/router": "^13.2.6",
+    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular-eslint/eslint-plugin": "^14.0.2",
+    "@angular/animations": "^14.1.2",
+    "@angular/cdk": "^14.1.2",
+    "@angular/cli": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/compiler": "^14.1.2",
+    "@angular/compiler-cli": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@angular/platform-browser": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.1.2",
+    "@angular/platform-server": "^14.1.2",
+    "@angular/router": "^14.1.2",
     "@fortawesome/fontawesome-free": "^5.15.3",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
