@@ -2,7 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { NotificationService } from '@service/notification/notification.service';
 import { HttpClient } from '@angular/common/http';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { startWith, map, exhaustMap, filter, tap, catchError } from 'rxjs/operators';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { Observable, of } from 'rxjs';
@@ -30,7 +30,7 @@ export class XParaphComponent implements OnInit {
     contextList = ['FON', 'PER', 'SPH', 'DIR', 'DLP', 'EXE'];
     addAccountMode: boolean = false;
 
-    usersCtrl = new FormControl();
+    usersCtrl = new UntypedFormControl();
     filteredUsers: Observable<any[]>;
 
     constructor(
