@@ -5,7 +5,7 @@ import { NotificationService } from '@service/notification/notification.service'
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { FunctionsService } from '@service/functions.service';
 import { tap, exhaustMap, map, startWith, catchError, finalize, filter } from 'rxjs/operators';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { LatinisePipe } from 'ngx-pipes';
 import { Observable, of } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
@@ -48,7 +48,7 @@ export class AvisWorkflowComponent implements OnInit {
     avisModelListNotLoaded: boolean = true;
     data: any;
 
-    searchAvisUser = new FormControl();
+    searchAvisUser = new UntypedFormControl();
 
     constructor(
         public translate: TranslateService,
