@@ -1,5 +1,5 @@
 import { Component, OnInit, EventEmitter, Output } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, ValidatorFn } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ValidatorFn } from '@angular/forms';
 import { NotificationService } from '@service/notification/notification.service';
 import { TranslateService } from '@ngx-translate/core';
 import { HttpClient } from '@angular/common/http';
@@ -16,11 +16,11 @@ export class DocserversComponent implements OnInit {
 
     @Output() nextStep = new EventEmitter<string>();
 
-    stepFormGroup: FormGroup;
+    stepFormGroup: UntypedFormGroup;
 
     constructor(
         public translate: TranslateService,
-        private _formBuilder: FormBuilder,
+        private _formBuilder: UntypedFormBuilder,
         private notify: NotificationService,
         public http: HttpClient,
         private installerService: InstallerService
