@@ -5,7 +5,7 @@ import { NotificationService } from '@service/notification/notification.service'
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { AppService } from '@service/app.service';
 import { SortPipe } from '../../../plugins/sorting.pipe';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { Observable, of } from 'rxjs';
 import { debounceTime, filter, distinctUntilChanged, tap, switchMap, exhaustMap, catchError, map } from 'rxjs/operators';
 import { LatinisePipe } from 'ngx-pipes';
@@ -28,7 +28,7 @@ export class TagInputComponent implements OnInit {
     /**
      * FormControl used when autocomplete is used in form and must be catched in a form control.
      */
-    @Input() control: FormControl;
+    @Input() control: UntypedFormControl;
 
     @Input() returnValue: 'id' | 'object' = 'id';
 
@@ -41,7 +41,7 @@ export class TagInputComponent implements OnInit {
     canAdd: boolean = false;
 
     listInfo: string;
-    myControl = new FormControl();
+    myControl = new UntypedFormControl();
     filteredOptions: Observable<string[]>;
     options: any;
     valuesToDisplay: any = {};
