@@ -5,7 +5,7 @@ import {
     Renderer2,
     Output
 } from '@angular/core';
-import { ControlValueAccessor, FormControl } from '@angular/forms';
+import { ControlValueAccessor, UntypedFormControl } from '@angular/forms';
 import { MatOption } from '@angular/material/core';
 import { MatSelect } from '@angular/material/select';
 import { take, takeUntil, startWith, map, debounceTime, filter, tap, switchMap, finalize } from 'rxjs/operators';
@@ -27,7 +27,7 @@ export class PluginSelectAutocompleteSearchComponent implements OnInit, OnDestro
     /** Label of the search placeholder */
     @Input() placeholderLabel = this.translate.instant('lang.chooseValue');
 
-    @Input() formControlSelect: FormControl = new FormControl();
+    @Input() formControlSelect: UntypedFormControl = new UntypedFormControl();
 
     @Input() datas: any = [];
 
@@ -85,7 +85,7 @@ export class PluginSelectAutocompleteSearchComponent implements OnInit, OnDestro
 
     @ViewChild('test', { static: true }) matSelect: MatSelect;
 
-    formControlSearch = new FormControl();
+    formControlSearch = new UntypedFormControl();
     noResult: boolean = null;
     loadingSearch: boolean = null;
 
